Coarsen working
There was quite a bit of really weird performance that I had to figure
out. Part of it was spending an enormous amount of time in verifying the
pqueue, but after I fixed that the Rust still performs way worse than
the original C version. I did some light profiling and didn't see
anything, so I'll have to come back and look more carefully.