port refine.c, also ran tests with ASAN
There was a buffer overflow being caused by a slice::fill since
get_graph_slices used graph.ncon * ctrl.npart as the size of ed slice,
but it was only ever allocated as 2 * ncon instead. Naturally this led
to issues (deferred of course). I only figured it out when I got the
bright idea to run with ASAN, which trivialized the problem.

Lessons learned:
- Use ASAN first if there might be memory bugs
- Don't try and run ASAN in gdb, export the environment variables to
  abort and coredump first, and then look at the core dump.