binary behavior appears to be identical
It turns out one of the issues is that at some point in the last couple
of version, Rust changed the algorithm behind sort_unstable. That proved
to be a nightmare to find.

Outside of that I fixed a number of smaller (particularly floating point
precision and text formatting) inconsistencies